+++ neutrino2/src/driver/gdi/color.h
@@ -129,8 +129,8 @@ extern "C"
 #define COL_BACKGROUND_PLUS_0			CFrameBuffer::getInstance()->realcolor[COL_BACKGROUND]
 
 ////
-int convertSetupColor2RGB(uint8_t r, uint8_t g, uint8_t b);
-int convertSetupAlpha2Alpha(uint8_t alpha);
+uint32_t convertSetupColor2RGB(uint8_t r, uint8_t g, uint8_t b);
+uint8_t convertSetupAlpha2Alpha(uint8_t alpha);
 uint32_t convertSetupColor2Color(uint8_t r, uint8_t g, uint8_t b, uint8_t alpha);
 
 ////
